Move malleable profiles (#131)
* move malleable profiles under listeners

* make agent id clickable from tasks table
