+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Turn on auto-refresh for the agent page by default
 -   Refactoring Agent Task components
 -   Add preview of Sponsor dashboard feature
+-   Move malleable profiles under listeners
+-   Make agent id clickable from tasks table
 
 ## [2.3.2] - 2023-05-22
